Changed to long instead of integer

+++ src/main/java/org/dannil/httpdownloader/repository/LoginRepository.java
@@ -3,7 +3,7 @@ package org.dannil.httpdownloader.repository;
 import org.dannil.httpdownloader.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoginRepository extends JpaRepository<User, Integer> {
+public interface LoginRepository extends JpaRepository<User, Long> {
 
 	public User findByPassword(String password);
